This commit was manufactured by cvs2git to create tag 'DataFormats- Common-V08-08-05'.
Sprout from master 2009-03-25 20:23:16 UTC William Tanenbaum <sha1-faefd172b4369b2002e8f0f7d11f304fd9896b5b@cern.ch> 'Add base class to Handles to reduce template bloat'
Cherrypick from master 2009-01-28 12:43:30 UTC Vincenzo Innocente <sha1-48872ea38f0e6361a52cfb8d57ac997c453089e2@cern.ch> 'compiles: added scope in typedef':
    interface/DetSetVectorNew.h
Delete:
    interface/OrphanHandleBase.h

+++ interface/DetSetVectorNew.h
@@ -165,7 +165,7 @@ namespace edmNew {
     };
     friend class FastFiller;
 
-    struct FindForDetSetVector : public std::binary_function<const edmNew::DetSetVector<T>&, unsigned int, const T*> {
+    struct FindForDetSetVector : public std::binary_function<const edmNew::DetSetVector<T>&, size_t, const T*> {
         typedef FindForDetSetVector self;
         typename self::result_type operator()(typename self::first_argument_type iContainer, typename self::second_argument_type iIndex) {
             return &(iContainer.m_data[iIndex]);
@@ -443,7 +443,7 @@ namespace edm {
     /* Probably this one is not that useful .... */
     namespace refhelper {
         template<typename T>
-            struct FindSetForNewDetSetVector : public std::binary_function<const edmNew::DetSetVector<T>&, unsigned int, edmNew::DetSet<T> > {
+            struct FindSetForNewDetSetVector : public std::binary_function<const edmNew::DetSetVector<T>&, size_t, edmNew::DetSet<T> > {
                 typedef FindSetForNewDetSetVector<T> self;
                 typename self::result_type operator()(typename self::first_argument_type iContainer, typename self::second_argument_type iIndex) {
                     return &(iContainer[iIndex]);
